getProperty: return `this` when no arguments
We encountered an
issue (https://github.com/iFixit/ifixit/issues/42675) where Matomo Tag
Manager expects a `get()` method to return an element if
present. Since MooTools causes `get` on an Element to pass calls
through to `getProperty`, we end up crashing because we call this
method with no arguments. An easy hack is to return the element in
that case, which doesn't require any other changes to the code, and
probably does exactly what the Tag Manager expects.

This might look like it wouldn't be safe, because it changes MooTools
behavior, but because it was incorrect to call `getProperty` without
an argument (previously `name.toLowerCase()` would crash if `name` was
undefined), there probably aren't a lot of places where this is
happening, and those that exist are probably bugs.